feat(broker): expose api catalog, topic-config docs, and FileConfig JSON schema
